+++ src/App.css
@@ -15,7 +15,7 @@
 }
 .app-wrapper{
   /* background-color: #64ab4418; */
-  min-width: 800px;
+  min-width: 750px;
   min-height: 600px;
   padding: 30px;
   box-sizing: border-box;
