--- src/components/Homepage1.css
@@ -1,40 +1,0 @@
-#row{
-    margin: 15px;
-    
-    display: flex;
-    justify-content: space-between;
-    
-}
-.container{
-    position:relative;
-    left: 10px;
-    /* background-color: ; */
-    
-    z-index: +1;
-}
-.row{
-    /* position: fixed; */
-    margin-top: 40px;
-    margin-left: 10px;
-    /* display: flex;
-    justify-content: space-between;
-     align-content: space-between; */
-} 
-.g-row-6{
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-    
-}
-#ee1{
-    background-color: aqua;
-}
-#fix{
-    position: fixed;
-}
-/* .abs{
-    position: absolute;
-} */
-
-
-
