+++ src/components/Sidebar.js
@@ -1,47 +1,28 @@
-import React, { useContext } from 'react';
+import React,{useContext} from 'react';
 import { userContext } from '../App'
 import Buttons from './Buttons';
 import Card from './Card';
 import Navbar from './Navbar';
 import Title from './Title';
-import "./Homepage1.css"
-
-function Homepage1({id,title,description}) {
+import "./Sidebar.css"
 
 
-  
+
+function Homepage({id}) {
   const setContext  = useContext(userContext)
   const{state,dispatch}= setContext
 
 
-
-  // console.log(setContext)
-
-
-console.log(id,'Hey i am card');
-
-// console.log(setContext.Todos)
   return (
-
-
-    <div >
-      
+    <div>
       {setContext.Open && <><div  className="container">
 
         <div className='app-wrapper '>
           <Navbar id={id} />
-          <Title id={id} />
-
+          <Title id={id} /> 
           <Buttons id={id} />
-      
-
         </div>
-        
-
-
       </div>
-      
-
       </>}
 
 
@@ -57,16 +38,29 @@ console.log(id,'Hey i am card');
         <div className='g-row-3'  ><h1 onClick={() => dispatch({ type: 'Open',id })} >+</h1>
         </div>
        
-
-
+      </div>
+    
+      <div className='col-sm-3 list'>
+        {/* <div id ='e2'> */}
+        <div id='a1'>
+          <div id='fcol1'></div><div>Study</div></div>
+          <div id='a2'>
+        <div id='fcol2'></div><div>Work</div></div>
+        <div id='a3'>
+        <div id='fcol3'></div><div>Entertainmnet</div></div>
+        <div id='a4'>
+        <div id='fcol4'></div><div>Family</div></div>
+        <div id='a5'>
+        <input id='fcol5' className='mt-5' type='checkbox'/>
+        <label className='ps-2'>Hide done tasks </label>
+        </div>
 
       </div>
+
+
     </div>
-
-
-
-
+  
   )
 }
 
-export default Homepage1
+export default Homepage
